Fix error in az skip check.
The actual az scans do no include a homing move, so when that is listed
as the most recent file, no az homing information was available. Work
around that by working backwards through files to catch a homing move.
This also makes the process robust against the case when no az homing
can be found in any recent file, and that is instead reported, too.